% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dinvWishart}
\alias{dinvWishart}
\title{Density of inverse Wishart random matrix}
\usage{
dinvWishart(x, df, S, log = FALSE)
}
\arguments{
\item{x}{array of dimension \code{d} by \code{d} by \code{n}}

\item{df}{degrees of freedom}

\item{S}{symmetric positive definite matrix of dimension \code{d} by \code{d}}

\item{log}{logical; if \code{TRUE}, returns the log density}
}
\value{
a vector of length \code{n} containing the log-density of the inverse Wishart.
}
\description{
Density of inverse Wishart random matrix
}
