% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dmbeta2}
\alias{dmbeta2}
\title{Matrix beta type II density function}
\usage{
dmbeta2(x, shape1, shape2, log = TRUE)
}
\arguments{
\item{x}{cube of dimension \code{d} by \code{d} by \code{n} containing the random matrix samples}

\item{shape1}{positive shape parameter, strictly larger than \eqn{(d-1)/2}.}

\item{shape2}{positive shape parameter, strictly larger than \eqn{(d-1)/2}.}

\item{log}{[logical] if \code{TRUE} (default), returns the log density.}
}
\value{
a vector of length \code{n}
}
\description{
Given a random matrix \code{x}, compute the density
for arguments \code{shape1} and \code{shape2}
}
