% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dsmlnorm}
\alias{dsmlnorm}
\title{Symmetric matrix-variate lognormal density}
\usage{
dsmlnorm(x, b, M, log = TRUE)
}
\arguments{
\item{x}{[cube] array of dimension \code{d} by \code{d} by \code{n}}

\item{b}{[numeric] scale parameter, strictly positive}

\item{M}{[matrix] location matrix, positive definite}

\item{log}{[logical] if \code{TRUE} (default), returns the log density}
}
\value{
a vector of length \code{n}
}
\description{
Density of the lognormal matrix-variate density, defined through the matrix logarithm, with the Jacobian resulting from the transformation
}
