% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dsmnorm}
\alias{dsmnorm}
\title{Symmetric matrix-variate normal density}
\usage{
dsmnorm(x, b, M, log = TRUE)
}
\arguments{
\item{x}{[cube] array of dimension \code{d} by \code{d} by \code{n}}

\item{b}{[numeric] scale parameter, strictly positive}

\item{M}{[matrix] location matrix, positive definite}

\item{log}{[logical] if \code{TRUE} (default), returns the log density}
}
\value{
a vector of length \code{n}
}
\description{
Symmetric matrix-variate normal density
}
