% Generated by roxygen2 (4.0.2): do not edit by hand
\name{generateLibrary}
\alias{generateLibrary}
\title{Generate kinase-subtrate mapping library}
\usage{
generateLibrary(path = ".")
}
\arguments{
\item{path}{character}
}
\description{
Generates library of kinase-substrate mapping data from raw Protein Prospector files.
}

