% Generated by roxygen2 (4.0.2): do not edit by hand
\name{generateUniprotMapping}
\alias{generateUniprotMapping}
\title{Generate mapping of Uniprot accession to protein sequence}
\usage{
generateUniprotMapping(path = ".")
}
\arguments{
\item{path}{character}
}
\value{
data.table
}
\description{
Generates mapping array from Uniprot accession to protein sequence from fasta files
available from http://www.uniprot.org/
}

