% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getExperimentSummary}
\alias{getExperimentSummary}
\title{Generate summary of kinase-substrate relationships for an experiment.}
\usage{
getExperimentSummary(exp, toFile = TRUE, path = ".", data = .dataFiltered)
}
\arguments{
\item{exp}{character}

\item{toFile}{boolean}

\item{path}{character}

\item{data}{data.table}
}
\description{
Description
}

