% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getGraph}
\alias{getGraph}
\title{Generate graph of substrates}
\usage{
getGraph(kin, org, interaction = "physical", go = TRUE,
  data = .dataAggregate)
}
\arguments{
\item{kin}{character}

\item{org}{character}

\item{interaction}{character}

\item{go}{boolean}

\item{data}{data.table}
}
\description{
Description
}

