% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getKinaseSummary}
\alias{getKinaseSummary}
\title{Generate summary of kinase-substrate relationships for a kinase.}
\usage{
getKinaseSummary(kin, toFile = TRUE, path = ".", data = .dataAggregate)
}
\arguments{
\item{kin}{character}

\item{toFile}{boolean}

\item{path}{character}

\item{data}{data.table}
}
\description{
Generate an aggregate table of kinase-substrate relationship for a given kinase.
Table is saved to file.
}

