% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getUniprotIDs}
\alias{getUniprotIDs}
\title{Get list of Uniprot IDs for kinase}
\usage{
getUniprotIDs(kin, data = .dataAggregate)
}
\arguments{
\item{kin}{character}

\item{data}{data.table}
}
\description{
Helper function to get list of Uniprot IDs of substrates of a given kinase.
Useful for downstream analysis such as GO annotation using DAVID (http://david.abcc.ncifcrf.gov/).
}

