% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lcv_kern_smlnorm}
\alias{lcv_kern_smlnorm}
\title{Likelihood cross validation criterion for symmetric matrix lognormal kernel}
\usage{
lcv_kern_smlnorm(x, b, h = 1L)
}
\arguments{
\item{x}{[cube] array of dimension \code{d} by \code{d} by \code{n}}

\item{b}{[numeric] scale parameter, strictly positive}

\item{h}{integer lag for excluding observations}
}
\value{
the value of the log objective function
}
\description{
Given a cube \code{x} and a bandwidth \code{b}, compute
the leave-one-out cross validation criterion by taking out a slice
and evaluating the kernel at the holdout value, excluding points that are at distance at least \code{h-1} apart.
}
