% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lscv_kdens_symmat}
\alias{lscv_kdens_symmat}
\title{Least square cross-validation for symmetric positive definite matrix kernels}
\usage{
lscv_kdens_symmat(x, b, h = 1L, kernel = "Wishart")
}
\arguments{
\item{x}{array of dimension \code{d} by \code{d} by \code{n}}

\item{b}{vector of candidate bandwidth, strictly positive}

\item{h}{integer for the lag vector for determining which observation to exclude, any data point in a radius of \code{h}}

\item{kernel}{string indicating the kernel, one of \code{Wishart} or \code{smlnorm}.}
}
\value{
a list with arguments
\itemize{
\item \code{lscv} vector of likelihood cross validation criterion
\item \code{b} vector of candidate bandwidth
\item \code{h} lag for leave-one-out
\item \code{bandwidth} optimal bandwidth among candidates
\item \code{kernel} string indicating the choice of kernel function
}
}
\description{
Given a cube of sample observations (consisting of random symmetric positive definite matrices), and a vector of candidate bandwidth parameters \code{b},
compute the least square likelihood cross-validation criterion and
return the bandwidth among the choices that minimizes the criterion.
}
