% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lscv_kern_smlnorm}
\alias{lscv_kern_smlnorm}
\title{Least square cross validation criterion for log symmetric matrix normal kernel}
\usage{
lscv_kern_smlnorm(x, b, h = 1L)
}
\arguments{
\item{x}{[cube] array of dimension \code{d} by \code{d} by \code{n}}

\item{b}{[numeric] scale parameter, strictly positive}

\item{h}{[int] integer indicating the separation lag}
}
\value{
a vector of length two containing the log of the summands
}
\description{
Finite sample h-block leave-one-out approximation to the least
square criterion, omitting constant term. Only pairs that are \eqn{|i-j| \leq h} apart are considered.
}
