% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lscv_kern_Wishart}
\alias{lscv_kern_Wishart}
\title{Least square cross validation criterion for Wishart kernel}
\usage{
lscv_kern_Wishart(x, b, h = 1L)
}
\arguments{
\item{x}{[cube] array of dimension \code{d} by \code{d} by \code{n}}

\item{b}{[numeric] scale parameter, strictly positive}

\item{h}{separation vector; only pairs that are \eqn{|i-j| \leq h} apart are considered}
}
\value{
a double containing the a vector of length two containing the log of the summands
}
\description{
Finite sample h-block leave-one-out approximation to the least
square criterion, omitting constant term.
}
