% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{meanlog}
\alias{meanlog}
\title{Log of mean with precision}
\usage{
meanlog(x)
}
\arguments{
\item{x}{vector of log components to add}
}
\value{
double with the log mean of elements
}
\description{
Log of mean with precision
}
\keyword{internal}
