% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rinvWishart}
\alias{rinvWishart}
\title{Random matrix generation from the inverse Wishart distribution}
\usage{
rinvWishart(n, df, S)
}
\arguments{
\item{n}{[integer] sample size}

\item{df}{[double] degrees of freedom, positive}

\item{S}{[matrix] a \code{d} by \code{d} positive definite scale matrix}
}
\value{
an array of dimension \code{d} by \code{d} by \code{n} containing the samples
}
\description{
Random matrix generation from the inverse Wishart distribution
}
