% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rmbeta2}
\alias{rmbeta2}
\title{Random matrix generation from matrix beta type II distribution}
\usage{
rmbeta2(n, d, shape1, shape2)
}
\arguments{
\item{n}{sample size}

\item{d}{dimension of the matrix}

\item{shape1}{positive shape parameter, strictly larger than \eqn{(d-1)/2}.}

\item{shape2}{positive shape parameter, strictly larger than \eqn{(d-1)/2}.}
}
\value{
a cube of dimension \code{d} by \code{d} by \code{n}
}
\description{
This function only supports the case of diagonal matrices
}
