% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rmnorm}
\alias{rmnorm}
\title{Random vector generation from the multivariate normal distribution}
\usage{
rmnorm(n, mean, vcov)
}
\arguments{
\item{n}{sample size}

\item{mean}{mean vector of length \code{d}}

\item{vcov}{a square positive definite covariance matrix, of the same dimension as \code{mean}.}
}
\value{
an \code{n} by \code{d} matrix of samples
}
\description{
Sampler derived using the eigendecomposition of the covariance
matrix \code{vcov}.
}
\examples{
rmnorm(n = 10, mean = c(0, 2), vcov = diag(2))
}
