% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rotation2d}
\alias{rotation2d}
\title{Rotation matrix for 2 dimensional problems}
\usage{
rotation2d(par)
}
\arguments{
\item{par}{double for angle in \eqn{[0,2\pi).}}
}
\value{
a 2 by 2 rotation matrix
}
\description{
Rotation matrix for 2 dimensional problems
}
\keyword{internal}
