% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rotation3d}
\alias{rotation3d}
\title{Rotation matrix for 3 dimensional problems}
\usage{
rotation3d(par)
}
\arguments{
\item{par}{vector of length 3 containing elements \eqn{\phi \in [0, 2*\pi)}, \eqn{\theta \in [0, \pi]} and \eqn{\psi \in [0, 2\pi)}, in that order.}
}
\value{
a 3 by 3 rotation matrix
}
\description{
Rotation matrix for 3 dimensional problems
}
\keyword{internal}
