% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rotation_scaling}
\alias{rotation_scaling}
\title{Rotation matrix with scaling for Monte Carlo integration}
\usage{
rotation_scaling(ang, scale)
}
\arguments{
\item{ang}{vector of length 1 containing  \eqn{\theta \in [0, 2\pi]} for
\eqn{d=2}, or of 3 containing elements \eqn{\phi \in [0, 2*\pi)}, \eqn{\theta \in [0, \pi]} and \eqn{\psi \in [0, 2\pi)} for \eqn{d=3}.}

\item{scale}{vector of length 2 (\eqn{d=2}) or 3 (\eqn{d=3}), strictly positive}
}
\value{
a 2 by 2, or 3 by 3 scaling matrix, depending on inputs
}
\description{
Rotation matrix with scaling for Monte Carlo integration
}
\keyword{internal}
