% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simu_fdens2d}
\alias{simu_fdens2d}
\title{Target densities for simulation study}
\usage{
simu_fdens2d(x, model)
}
\arguments{
\item{x}{cube of dimension \code{d} by \code{d} by \code{n} containing the sample matrices}

\item{model}{integer between 1 and 6 indicating the simulation scenario}
}
\value{
a vector of length \code{n} containing the density
}
\description{
Target densities for simulation study
}
\keyword{internal}
