% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simu_rdens}
\alias{simu_rdens}
\title{Target densities for simulation study}
\usage{
simu_rdens(n, model, d)
}
\arguments{
\item{n}{sample size}

\item{model}{integer between 1 and 6 indicating the simulation scenario}

\item{d}{dimension of the matrix, an integer between 2 and 10}
}
\value{
a cube of dimension \code{d} by \code{d} by \code{n} containing the sample matrices
}
\description{
Target densities for simulation study
}
\keyword{internal}
