% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sumlog}
\alias{sumlog}
\title{Log of sum with precision}
\usage{
sumlog(x)
}
\arguments{
\item{x}{vector of log components to add}
}
\value{
double with the log sum of elements
}
\description{
Log of sum with precision
}
\keyword{internal}
