% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sumsignedlog}
\alias{sumsignedlog}
\title{Signed sum with precision}
\usage{
sumsignedlog(x, sgn)
}
\arguments{
\item{x}{vector of log components}

\item{sgn}{sign of the operator}
}
\value{
log sum of elements
}
\description{
Given a vector of signs and log arguments,
return their sum with precision avoiding
numerical underflow assuming that the sum is strictly positive.
}
\keyword{internal}
