% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{symmetrize}
\alias{symmetrize}
\title{Symmetrize matrix}
\usage{
symmetrize(A)
}
\arguments{
\item{A}{square matrix}
}
\value{
symmetrized version of \code{A}
}
\description{
Given an input matrix, symmetrize by taking average of lower and upper triangular components as \eqn{A + A^\top}.
}
