% Generated by roxygen2 (4.0.1): do not edit by hand
\name{generateUniprotMapping}
\alias{generateUniprotMapping}
\title{Generate mapping of Uniprot accession to protein sequence}
\usage{
generateUniprotMapping(path = ".")
}
\arguments{
\item{path}{character}
}
\value{
data.table
}
\description{
Description
}

