% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getConsensus}
\alias{getConsensus}
\title{Generate list of local sequences around phosphorylation site.}
\usage{
getConsensus(kin, ids = 1, toFile = TRUE, path = ".",
  data = .dataAggregate)
}
\arguments{
\item{kin}{character}

\item{ids}{integer}

\item{toFile}{boolean}

\item{path}{character}

\item{data}{data.table}
}
\description{
Generate list of aligned sequences for determination of consensus phosphorylation sites
using a web-based interface such as Weblogo3 (http://weblogo.threeplusone.com/create.cgi).
}

